Adsb Maven Refactor
I have refactored the code to work with the maven package manager. This will allow the ease of installing and using testing frameworks and libraries in the future, starting with JUnit. I also made some changes to the code for things that didn't make sense and also temporarily disabling the GUI version until I get the algorithm down.